src/requirements.txt
absl-py==0.2.0
affine==2.2.0
astor==0.6.2
attrs==17.4.0
bleach==1.5.0
click==6.7
click-plugins==1.0.3
cligj==0.4.0
cycler==0.10.0
Fiona==1.7.11.post1
gast==0.2.0
GDAL==2.2.4
grpcio==1.11.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.6
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
munch==2.2.0
numpy==1.14.2
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
rasterio==1.0a12
rasterstats==0.12.1
rope==0.10.7
scikit-learn==0.19.1
scipy==1.0.1
Shapely==1.6.4.post1
simplejson==3.13.2
six==1.11.0
sklearn==0.0
snuggs==1.4.1
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
